Add support for --chown <gid>:<uid>
The parameter can be given in the form <uid>:<gid> but also username and
group names are supported _if_ the os supports the lookup.
The safest way is to use numeric user ids.

The use case for this feature is when the processed templates are expected
to belong to a certain user/group. E.g. Wildfly Docker images, a prominent Java
EE application server write into the configuration files, too but runs with a non-root uid
and so needs the proper access. Depending on the uid under which the server
run and the files has been created, this has to be tuned.

Especially in a Docker context this might be helpful since Docker does not
support changing the ownership with ADD or COPY (its always added as root).
In order to avoid some postprocessing of the generated files (which might even
not be possible e.g. when running from a scratch image without any installed
shell), this options proves to be quite useful and add some flexibility
(on the cost of an additional feature, though).