Document suppressing newlines. (#450)
* Document suppressing newlines.

I found myself writing multi-line action blocks to improve readability but could not understand why each action left a newline behind. Because I was running this in Docker Compose, I didn't want to alter the `ENTRYPOINT` to your image to do some data munging with `bash` to workaround it.

I saw that some example code in your documentation had leading and trailing minus signs after and before the action delimiters. I figured this was worth something, so I used it in my code and discovered that this is how you suppress newlines.

While this is well-known for those that use Go templates, it wasn't clear to this not-yet-a-Golang dev. I added some documentation to explain this better for folks like me.

I love this tool! Thank you for making it!

* Update docs/content/syntax.md

Co-Authored-By: carlosonunez <13461447+carlosonunez@users.noreply.github.com>

* Update syntax.md

Reference Golang template behaviour.

* Apply suggestions from code review

Co-Authored-By: carlosonunez <13461447+carlosonunez@users.noreply.github.com>