Update syntax.md
fix: wording was incorrect.

Details: Original text stated
```... newlines that come before the action) can be suppressed by placing a minus sign in front of the first set of delimiters (`{{`)...```

This appears to have been a typo. Following those directions would lead to (`-{{`). This would not behave as expected. What was originally intended was (`{{-`)